Updated 3.6.1
Save Image Click Settings - When you added an image click to a 3D Web Object - Image, it was not saving the jsfunction and jsparameters to the database. Turns out that the values were being passed as imagejsfunction and imagejsparameters while being read as jsfunction and jsparameters. We changed the values to jsfunction and jsparameters. Keep in mind that if the mold needs to use these values, we may need to modify this in the future. The Mold and the Web Image tables both have jsfunction and jsparameters.

Image Click Parameter Correction - When a 3D Web Object - Image is clicked, the checkImageClick function was sending the moldname as a second parameter to the setFunctionAndExecute function where it was combining it with the base parameter for execution. We removed the moldname form the passed parameters and it now works as expected.

Immediately Show Image on Change - When editing a 3D Web Object - Image, when you change the image, it was not updating. The setNewMold function needed to include the path update for the image (was previously only imageid based). (Reference: /core/scripts/admin/wtw_adminmolds.js setNewMold function)

Image and Hover Image Path Support - On the molds form, when editing a 3D Web Object - image, the form thumbnail was not showing the image. Previously the image data was stored in the database, now it is on the file system and uses a path reference. The code to open the form now supports the proper path methods. (reference: /core/scripts/admin/wtw_adminmolds.js openMoldForm function)

Default New Image Rotation - When adding a 3D Web Object - Image to a 3D Scene, the default rotation was off by 90 degrees. This has been corrected in the /core/scripts/molds/wtw_addmoldlist.js file.

Coverings List for Plugins values correction - When a new covering is added to the list from a plugin, the values in the array were strings, while the plugins were returning an object format. For now we are reverting to the string value. This may change in the future as the plugin use of coverings becomes more expansive.

Coverings List for Plugins - When entering a new covering from a plugin, the list was using a local variable, but the returned value from the plugins was assigning a non-existent global. The global variable has been replaced by the local variable.

Parameter Removed from Covering List - The function returnCoveringDefsList in the class plugins, does not have a passed parameter, yet the call to the function was trying to pass the value as covering. The parameter has been removed.

Updated Global Login Settings - Since we renamed the login settings settingname(s), we added code to rename the settings in the database to preserve the settings values.