Updated 3.6.1
Added Note for First 3D Building - We added a note instruction to the First 3D Building page under 3D Community -> Options and Settings. This setting is only found when editing a 3D Community. It is used by Wizards and Automated installs to allow the user to select a 3D Building and 3D Community and automatically have that building added to the community scene on completion. This is that location position, scaling, and rotation that is used.

3D Model Collisions - We added code to overwrite the obj, glb, gltf, and babylon 3D Model child meshes check collisions setting. When you set the full Mold to check collisions in the advanced options while editing the mold, when it is set to true, it will automatically add check collisions. When it is set to false, it will merely inherit the original settings. (that way you can have some meshes with collisions and others without if done on the design level). Note the default is to have check collisions set to true.

Image Collisions Corrected - With the new Done covering setting, we were able to use the newly passed checkcollisions value to turn off or on collisions on the main box that holds the images. This corrects the images and raised images collisions issue.

Covering Done setting - When adding Coverings to Molds, there are now 2 key words: none and done. None is used by things like 3D Models that do not require any further actions for materials and collisions (note that the collisions are handled on each sub mesh separately). Done is used to not add materials, but still allow the main collisions and isPickable flags to be set on the main mesh.

Setting for Collisions and Pickable - When editing a mold, the form now has entries for Check Collisions and is Pickable under the Advanced Settings. They will default to true if not set. Note that all items are pickable in Admin Mode; this setting is to show a hand cursor in Browse Mode.

Check Collisions and isPickable are No Longer Defaults - The Check Collisions and isPickable settings for a mold are no longer just set as 1 (true). The call to the database, handlers, classes for molds to store and retrieve the settings is now added.