Updated 3.6.1
Coverings List for Plugins values correction - When a new covering is added to the list from a plugin, the values in the array were strings, while the plugins were returning an object format. For now we are reverting to the string value. This may change in the future as the plugin use of coverings becomes more expansive.

Coverings List for Plugins - When entering a new covering from a plugin, the list was using a local variable, but the returned value from the plugins was assigning a non-existent global. The global variable has been replaced by the local variable.

Parameter Removed from Covering List - The function returnCoveringDefsList in the class plugins, does not have a passed parameter, yet the call to the function was trying to pass the value as covering. The parameter has been removed.