Updated 3.6.1
Created Hook for pluginsGetPluginInfoComplete - In an effort to isolate the Global references to the 3D Internet and 3D Avatar Plugins, we added a JavaScript hook for pluginsGetPluginInfoComplete. We then moved the related code to a new function called getPluginInfoComplete under the 3D Internet Plugin (with the necessary PHP add script call and Plugin Class script entry). This hook allows the 3D plugins to add code to the returned plugin list array and display. For the 3D Internet Plugin, it allows us to check for updates and add download buttons for updates as well as badges.

Moved plugin Hook for pluginsLoadLoginSettings - The plugin hook for pluginsLoadLoginSettings has been moved from the loadLoginSettings to the loadLoginSettingsResponse function so that it can use the WTW.globalLogins settings in the plugins.

Added Enter Button Code - We added code to show the enter button on the loadUserSettingsAfterEngine function. We placed it right before the plugins hook, so that plugins can override or reset it as necessary.

3D Internet Plugin Activation - During the new install, if you select default installation, the 3D Internet Plugin will not automatically load as activated. The 3D Internet Plugin code is still there and can easily be activated in the Admin -> 3D Plugins. If you select custom installation, the 3D Internet Plugin is automatically set active and you will have options to download your 3D Building and 3D Community selections in the install.

Place Holder Camera Movements - We added a section in the move avatar function to handle Camera Movements when the avatar is not loaded yet. This was required as a separate section since the core avatar movements is a weighted movement based on animations running and there are no animations running on the place holder before the avatar is selected.

Place Holder Limitations Removed - There were a number of inputs and move avatar commands that sere set to only work if the placeholder was removed and the avatar was present. We removed these restrictions so that the initial camera can move before the avatar is started on the scene. (WTW.placeHolder == 0)

3D Scenes start with Collapsed Menu - The 3D Scenes now load with a collapsed browse menu. Click the bottom left icon to toggle the browse menu open or closed.

Enter Scene Camera Movement Enabled - When you enter a 3D Scene, we have now enabled a limited amount of camera movement. Keyboard, mouse, or touch will now allow you to rotate the camera right, left, up, and down. This allows you to look at a 3D Scene before you choose to enter it, login, select an avatar, and walk around.

New Enter Button on 3D Scene Load - When you first load a new 3D Scene, you will now see a single small button labeled ENTER. When you click it, you will then see the login options (or straight to a login if you have only one option enabled).