Updated 3.6.1
Feedback Notes added - We added some notes to the feedback page under Admin Dashboard. The note provides instructions where users submit the feedback, and we added a note for when no feedback is found.

API Keys Note - We added a note to indicate when there are no API Leys found. If found, the normal list will show.

Developers Tools Note - We added an Admin Menu -> developers tools note to instruct to click F12 to open the console to display the following lists.

Created Hook for pluginsOpenFullPageFormMediaLibrary - In an effort to isolate the Global references to the 3D Internet and 3D Avatar Plugins, we added a JavaScript hook for pluginsOpenFullPageFormMediaLibrary. We then moved the related code to a new function called openFullPageFormMediaLibrary under the 3D Internet Plugin (with the necessary PHP add script call and Plugin Class script entry). This hook allows the 3D plugins to add code to media library section of the open full page form function.

The Update Menu Item Code Moved - The Update menu item under the Admin Dashboard section was moved to the 3D Internet Plugin, since the page it opens was already moved.

Swapped Columns on Dashboard - We swapped the columns on the Admin Dashboard because the first column was all provided by the 3D Internet Plugin and when it is inactive the first column was blank. Now it looks more appropriate.

Google Translate Prompt - We were getting a Google Translate bar opened to Arabic when opening the Admin Mode. We added a meta tag to set google to notranslate and a tag to identify the html tag as English language.

Enable Plugin Error - When adding the enable 3D Internet Plugin code if you select custom install, the code was missing a closing quote and semicolon. This was causing a 500 error in the install. This has been corrected.

Toggle Admin Menu Items - In the toggle admin menu items code, we check if any sub menu items are already open, then close on toggle. We moved 2 of the items to the 3D Internet Plugin, so code needed to test if the plugin was enabled and the options are set.

Admin Menu Upgrades Icon - The Upgrades admin menu item was missing the icon (it was in a plugin that is not released yet.) We moved the icon to the content/system/images folder and corrected the paths in the code.

My Avatar Mobile Transparency - When using mobile devices, we set the transparency to 50% in order to increase the visibility of the 3D Scene.

Render Canvas Style - After the latest Babylon update, the render canvas was defaulting to show the pointer cursor all of the time. We set it manually to cursor at default to over-ride this setting. Now the mouse changes to pointer on pickable elements in the 3D Scene.

Centered Avatar Animations Selections - After collapsing the browse menu, the avatar animations would pop up on the left side when opened. We changed the centering calculation to be WTW.sizeX based to correct the issue.