Updated 3.6.2
New Has Input Moved function - When testing the difference between a swipe and a click, we check to see if the mouse x or mouse y movement has changed. But under some circumstances event the smallest movement was not counting as a click. This function replaces the (WTW.mouseX == zstartx && WTW.mouseY == zstarty) comparison to accept a little movement as still being a click. The zdist setting is the tolerance. Now most touch clicks and mouse clicks are caught and processed correctly.

Added Input Moving function - We added an inputMoving function to the /core/scripts/prime/wtw_input.js so that the mouseMove and the touchMoving functions can share the same common code.

Added Input Click function - We added an inputClick function to the /core/scripts/prime/wtw_input.js so that the mouseClick and the touchClick functions can share the same common code.

Added Input Up function - We added an inputUp function to the /core/scripts/prime/wtw_input.js so that the mouseUp and the touchUp functions can share the same common code.

Added Input Down function - We added an inputDown function to the /core/scripts/prime/wtw_input.js so that the mouseDown and the touchDown functions can share the same common code.

Added Boolean Response to checkMoldFunctionAndExecute - We added a Boolean response to the checkMoldFunctionAndExecute function to determine if a function was found and executed. This helps determine if the click event found a mold with a function or needs to allow something else to still execute, such as Avatar Movement based on the click location.

Eliminate the Obsolete itouch Divs - We eliminated the obsolete itouch divs (wtw_itouchleft and wtw_itouchright). They were from an earlier version of touch movements that is no longer valid. All references and styles related to it have also been removed.

Adjusted the Camera Rotation Max - We adjusted the camera up and down rotation limits to shorten the distance of rotation. These settings can be found in the /core/scripts/avatars/wtw_avatarfunctions.js file moveAvatar function.

Renamed WTW.testTimer to WTW.avatarTimer - We renamed and repurposed the global WTW.test timer to be names WTW.avatarTimer and is used to test or execute automated avatar movement commands when they require an increment timer.

Open HUD Login for Enter - We set up specific conditions to show the Enter button from the Login Menu HUD. This allows us to easily pull that Enter button on the screen without the rest of the Login Menu showing.

Connect Avatars now Accepts Additional Groups - With the new Additional Avatar Groups allowing the same avatar to exist in multiple groups, we modified the SQL in the avatars.php connect file to add the avatar group selection in the where clause to include the avatarsingroups and avatargroups tables.

Random Selection of Avatar on Enter - We changed the process to start with a random avatar selection when you click Enter to join a scene. This process uses any avatar that you set to your Anonymous Group. It can be the primary Avatar Group or additional Avatar Group.

Save Avatars in Groups Functions - We added a saveAvatarsInGroup function in the avatars class to handle changes to the avatars in groups table. It uses the avatars.php handler.

Added Additional Avatar Groups Section - When editing an Avatar, under Edit 3D Avatar - 3D Avatar Information, we now have 2 parts to the avatar groups. The first is the primary avatar group that is used when shared. The second is a list of all possible groups from the avatar groups table allowing you to select as many as you wish. This will add the avatar into the group(s) for when they are used on a 3D Community - in the avatar selection to enter.

Added Avatar Groups to Responses - We added the avatargroups to various connect call responses.

Connect Avatars added Groups - We added the pulled avatar avatargroups to be returned with the connect avatars.php file.

Connect Avatar added Groups - We added the pulled avatar avatargroups to be returned with the connect avatar.php file.

Populate the Avatars in Groups Table - We added code to the checkDBVersionData function to populate the new avatarsingroups table based on the avatargroup field in avatars table and the avatargroupid found in the avatargroups table using that avatargroup name.

Added Table for Avatars in Groups - The hosted Avatars that users can select from to create their user avatars only had one avatar group. Now we are adding a table so that we can make it a one to many relationship and one avatar can be selected from multiple groups. We created avatarsingroups table.

Community Ground Naming Update - The input page processes clicks on the 3D Objects marked as pickable and have a hyphen in the name (meaning they are fully named objects). The Extended Ground in the scene had a simple name and was designed to be ignored, but now that we want to add the mouse and touch movement when you click on the ground, we needed to add a hyphen to the name (simplest solution was add a hyphen at the end). So the communityeground is now communityeground-.

Add 3D Building to Community Corrected - After the move of the 3D Internet Plugin code last month, a handful of the admin menu switch case items were moved when they should have been split to both the main site and 3D Internet Plugin. While the Forms and Menu Items are there, the case that selects the code to run was moved and will be inactive if the 3D Internet Plugin is deactivated. There are 2 files for this update and can be downloaded from https://github.com/HTTP3D/WalkTheWeb. The files are: /core/scripts/admin/wtw_adminmenus.js and /content/plugins/wtw-3dinternet/scripts/class_main.js