Updated 3.6.2
Removed Avatar Select Functions - The following functions from wtw_login.js have been removed as obsolete: getMyAvatarList, getFullAvatarList, getAnonymousAvatarList, and getMyAvatarList. These functions have all been replaced by the new Select My Avatar Form and functions.

Removed pluginsGetMyAvatarList Plugin Hook - The plugin hook called pluginsGetMyAvatarList has been removed. The process has become obsolete with the new 3D Form for Select My Avatar.

Added Plugin Hook for pluginsHudLoginLoadAvatarsArray - We added a hook for pluginsHudLoginLoadAvatarsArray that is fired after the initial Local Login Avatars are loaded to the Select My Avatar form. This plug allows additional code to execute, allowing additional avatars to be processed or loaded to the form. For the 3D Internet Plugin, this adds the WalkTheWeb Global Avatars to be added to the selection array.

New Select My Avatar 3D Form - We added a new form for Select My Avatar that replaces the HTML version. This new form works in the 3D Scene and will also work in VR cameras.

Click on items to Cancel Walk to Position - You can now click on an item or My Avatar to stop an in progress Walk To Position.

Animation Icon Moved Folders - While we no longer used the icons folder under content/system, there is one default image used. We moved animdefault.png from the icons folder to the images and updated the path in code.

Profile Icon now Toggles - We set the icon for Login on the Browse Menu to toggle the login menu or user menu when it is clicked. This makes it easier to close the menu.

Corrected new install avatar selection - During a new install, the user is logged in, but the avatar is not selected. We needed to adjust the code to handle this new condition. Now the avatar is randomly selected unless the user chooses to select an avatar.

Updated Map Files - Updated Map files to not include source (can be overwritten by the WalkTheWeb Developers Toolkit map files)