Updated 3.6.2
Limited Movement during Place Holder - When you first enter a 3D Scene and you can see the Enter button, we removed the movement for forward and backwards since the placeholder is designed to not move until you actually enter the 3D Scene. This fixed the camera issue that was showing an endless dive falling into the ground. We also disabled any click movement during this same place holder time.