Updated 3.6.2
Swipe Movement Function - Since the swipe direction changes as you move your avatar, we added a separate swipe function to determine the movement added to the avatar specifically for rotation left, right, up and down. These are some experimental settings. For left and right, the 1/5 sides of the screen will rotate in a run mode (faster) while the closer to center moves in a walk mode (slower) for rotation, and the center will not rotate left or right. For up and down, Movement to the top or bottom will look up and down, while near the center will move toward centering the camera once again.

Interval Timer added to Input Moving - While you hold the mouse down or the touch down, we added an interval timer so that the key pressed continues to activate until you release the mouse or touch. Previously, it was only sending the key pressed while you were in movement.

Mouse Timestamp and Clicks - We added a new global variable WTW.mouseTimestamp that is used to determine if the mouse down to mouse up time is less than 350 ms and should considered a click. This keeps the click event from processing unless it is just a click with no hold. This same timer process also catches the difference in a touch and hold vs a touch and release.

Disable Side Swipe Refresh on Mobile - In order to improve the swiping ability on Mobile devices, we disabled the Side Swipe page refresh by containing the over scroll behavior for the x direction in the stylesheet.

Portrait View Avatar Visibility Fade - We already made My Avatar 50% visibility when on a mobile device, now we expanded that visibility setting for any time the device is in portrait view by testing if the size X is less than the size Y.

Disable pull-down Refresh on Mobile - In order to improve the swiping ability on Mobile devices, we disabled the pull-down page refresh using position zero and containing the over scroll behavior in the stylesheet.