Updated 3.6.2
Added Animations, Hovers, and Click Functions - We added animations, hovers, and click functions for the 3D Form close buttons. This includes the references to the materials for color changes. All but one of the close buttons are named button-close and button-closetext. The remaining button is called button-closeloginmenu and button-closeloginmenutext since that form has a different pattern of visibility based on when the Enter Button alone is shown.

Hidden Form Fields are now Changed - The hidden form fields used to feed the 3D Text Fields are no longer the type hidden, now are type text and are hidden by moving them to an absolute position above the visible screen. This allows the text box to receive focus and open the keyboard on mobile devices.

Added Close Buttons to 3D Forms - To make it easy to cancel a 3D Form, we added an animated 3D Button to close the 3D Form on all of the 3D Forms (Login Menu, Login, Create Account, Recover Password, User Menu, and Select My Avatar).

Added Direction to Tab Next Field - We added a direction variable to the tab next field function (tabNextField(zdirection) that will now process a negative direction as going backwards through the tab order array.

Added Global Text Cursor - We added a global variable for the text cursor position: WTW.textCursor. This is used to mark the cursor position within a field value.

Added Cursor Functionality - We added cursor functionality. You can now use left and right arrow keys to move the cursor, shift-Tab or up arrow will move the focus to the previous input element, Tab or down arrow will move the focus to the next input element, del will remove the letter in front of the cursor, and backspace will remove the letter before the cursor.