Updated 3.6.3
New Ratings Menu Icon - We also added a menu icon to the Content Ratings item in the mobile menu. This change also updates the menu items table population to include the icon path as well as the table data update to the record for the path. The icon is not shown in the traditional browse menu.

New Mobile Browse Menu - We created a new Mobile Browse Menu. This is a copy and reformatted version of the bottom browse menu. The code changed in the class_wtwmenus.php, core stylesheet, and scripts in the prime, avatar, admin, and hud folders. Every place that updated the browse menu, now updates the mobile menu also.