Updated 3.6.3
Cancel Movement with Cancel Walk To Position - When the user clicks or touches the screen, it places a mark on the ground and the avatar will walk to that position. We added code that when the user clicks the keyboard, drags the mouse, swipes the screen, or moves the scrollbar, the click to movement will stop. This update is in the cancelWalkToPosition function and called by a number of input functions.

Enter Button White Text Corrected - The Enter Button text did not have a material assigned to it. So it would appear as the default gray. Then when you hovered over it, it would change colors, and then finally change to the correct white color. Now the 3D Object s updated to correct the default text color. In addition, some of the other 3D Objects on the login menu had unused textures. Now they have been removed.