Updated 3.6.3
New Mouse Out function for Molds - Using the pattern of the new mouse over by render group, this mouseOutByRenderingGroup function is called to clear the mouse overs and serves as a mouse out function.

New Mouse Over By Render Group - We created a new function called mouseOverByRenderingGroup that replaces the WTW.mouseOver action manager controlled function. This new function respects rendering group ids and creates a new mouse over and mouse out events.

New Pick Mold Name By Rendering Group - We created a new function called pickMoldNameByRenderingGroup in the wtw_inputs.js file. This function uses a pick ray to select multiple meshes under the cursor, then parses through them to select the highest render group with the shortest distance. This allows us to select buttons on the HUD when the render is showing it first regardless of closer 3D Objects. (Previously the render was working to order the 3D Objects, but the select mesh was not.)

Added Pick Mold Name Function - Since numerous inputs use the pick mold name based on mouse or touch positions, we moved the common code to a function called pickMoldName(zevent). This will allow us to enhance the function as needed in one place.

Input Up Modified - We moved the mouseUp code for Drag ID and clauses to the inputUp function so that it will now work for touchUp also. In addition, the cancelWalkToPosition is present, it was redundant to include the same keyPressedRemove commands so they were removed. Also note that the Input Up fires before the click or long click, so they do not interfere with the click movement being initiated.

Input Moving Corrected - When you clicked on the screen sides the avatar would jump slightly towards that touch. We corrected this action by confirming that the mouse or touch has moved positions before triggering the motion. Now it performs as expected and the avatar will not turn unless you initiate a mouse or touch movement (or keyboard commands).

Added Long Click Event - When a click is held longer than 350 ms and the mouse or touch has not moved on the screen, it is now triggering a long click event (inputLongClick function added to the wtw_input.js file).

Slow Click for Avatar Animations - We the ability to slow click on your avatar to open the Avatar Animations selection menu. Optionally, slow click again to close it.

Cancel Movement with Cancel Walk To Position - When the user clicks or touches the screen, it places a mark on the ground and the avatar will walk to that position. We added code that when the user clicks the keyboard, drags the mouse, swipes the screen, or moves the scrollbar, the click to movement will stop. This update is in the cancelWalkToPosition function and called by a number of input functions.

Enter Button White Text Corrected - The Enter Button text did not have a material assigned to it. So it would appear as the default gray. Then when you hovered over it, it would change colors, and then finally change to the correct white color. Now the 3D Object s updated to correct the default text color. In addition, some of the other 3D Objects on the login menu had unused textures. Now they have been removed.