Updated 3.6.3
Select Avatar Save Time Reduced - When changing avatar using the select avatar 3D Form, if you are using a global login, the save button takes a long time to copy your chosen avatar to the global server. Previously, the avatar would not load to the scene until the process is finished. Now, we are loading the local version being copied immediately so you can visit the scene and the copy process continues in the background.

User 3D Menu Buttons Renamed - The names of the buttons had myavatar in them, so when we excluded myavatar so it can click through your own avatar when selecting, it was also ignoring the buttons. They are now renamed and working properly. Also, to select you own avatar, use the long click.

Enter Button Resolved - If you panned the view before clicking enter, the enter button would animate but not activate to enter the scene. This has been resolved by changing the if clause conditions in the inputClick function and removing the else part of the statement.

Mouse Over and Out Global Variables Removed - The WTW.mouseOver and WTW.mouseOut global variables are obsolete so have been removed from the code. They were used with the register mouse over and action Manager obsolete items that were also removed.

Register Mouse Over Function Removed - The registerMouseOver function is obsolete so has been removed from the code. This includes the 102 reference calls to it throughout the code.

Mouse Over Action Manager Removed - We removed the Shopping Plugin moldQueueAdd function because it did not account for the rendering group order when hovering over 3D Objects in the scene. This functionality has been replaced with the mouseOverByRenderingGroup function in wtw_input.js

Mouse Over Action Manager Removed - We removed the Mouse Over and Mouse Out Action Manager because it did not account for the rendering group order when hovering over 3D Objects in the scene. This has been replaced with the mouseOverByRenderingGroup function in wtw_input.js

Mouse Out Function Removed - The mouseOutMold function has been replaced by the mouseOutByRenderingGroup function. Therefore obsolete and removed from code.

Mouse Over Mold Function Removed - The mouseOverMold function has been replaced by the mouseOverByRenderingGroup function. Therefore obsolete and removed from code.