Updated 3.6.3
Get My Images function Extended - We modified the getMyImages php function SQL to include a case for the new standard files category. It now returns documents and skybox files.

Media Library Files Icon - We had to add code for the skybox dds, hdr, and exr types of images to not try to show a thumbnail since they are not standard image types. The code for the media Library and Information about the file now has these exceptions and replaces the thumbnail with a simple file icon.

Babylon v6.1.0 - We upgraded the Babylon Engine for the v6.x.x folder to version v6.1.0.

Added Standard File as Media Library Type - We needed a new category for files that are not documents, images, videos, or audio files, so we added a standard Files category. This will host anything that does not fit in other categories. Files category will also show the documents category items. But it will also include things like dds and hdr SkyBoxes that are not standard images that can easily show a thumbnail.

Physics Engine Selection - In Admin Mode under Settings, starting with version v6.x.x, you now have the ability to select the physics engine of your choice. Previously, cannon engine was enabled and the cannon and oimo engines were being loaded by default (which still works if you select v5.x.x). v6.x.x now includes the Havok Physics engine.

Select Babylon Version - In the Admin Mode, Settings, you will see a new section that includes a drop-down for Babylon Version. This allows you to select which version you wish to operate on your server (as a whole server).

Multiple Babylon Versions Supported - Since some of the new versions may need to be tested before you can update your site to the new version, we have added a folder structure and selection in code that will allow multiple versions to exist under the /core/scripts/engine folder.

Merge Files for htaccess updates - In the class_wtw-initsession.php file, we added a function to mergeFiles that we are using with the htaccess file. GitHub does not show .htaccess files so we use a htaccess file without the dot to represent what should be there. During a new install, the file is fully copied to the .htaccess filename. After initial install, we know some changes and customizations may exist that we respect and preserve. So we created a mergeFile function that will take every line in our htaccess file and look to see it is in the .htaccess file. If not, it will be added automatically. For best results, we recommend that any changes you make are before the WalkTheWeb section.

New Mime Types Supported - We added 4 new mime types: wasm is used by the new havok physics engine and the other three (dds, hdr, and exr) are for sky boxes. The dds and hdr files are already in use for the PBR SkyBox and HDR SkyBox respectively. The exr is for future use. These same mime types are added to the web.config for Windows Servers.

Physics Engine Global Variable - We eliminated the WTW.enablePhysics and replaced it with WTW.physicsEngine that accepts values of havok, cannon, oimo, or blank for none.

Added Mime Types to htaccess - We added some mime types to the htaccess file to now include: dds, hdr, exr, and wasm. dds and hdr can be used for skyboxes, exr is for future use, and the wasm file was added for usage by the Havok Physics Engine.

Added Function to Get Max Upload Size - We added a function to get the maximum upload file size value in bytes. We also added a function to convert values with P, T, G, M, or K to bytes to assist the first function. These functions are also extended to handler, connect, and plugin classes. It is also now used to replace the 5 cases of 128000000 hard coded value.

Physics Viewer Added to Admin Mode - We implemented the Physics Viewer. You will find it in Admin Mode, under Development Tools. This tool will allow you to see how Babylon implements the physics around objects and help troubleshoot physics related implementations.

Babylon Debug Functionality Added - We implemented the Debug Layer Inspector. You will find it in Admin Mode, under Development Tools. This tool will allow you to see how Babylon implements meshes and functionality in the scene.

Beginning adding Physics to Molds - We are beginning to add physics to molds. This will continue to be implemented for a couple months. You can expect physics behaviors to improve as we go.

Updated BabylonJS to 6.0 - We updated BabylonJS to version 6.0! Read about all of the new features at: https://babylonjs.medium.com/announcing-babylon-js-6-0-dcb5f1662e3a

We have implemented Havok Physics Engine - We have implemented the Havok Physics Engine! This was released in the BabylonJS version 6.0.

Split Load Sequence in two Functions - We split the load sequence in the wtw_core.js file into two parts (second is continueLoadSequence) so that the physics engine can be loaded async and the remainder of the load sequence can be triggered on its completion. Some of the items cannot be loaded until the scene is created properly with the physics engine loaded.

Added Set Sky Size Function - We added a setSkySize function so that the size of the sky could be changed without reloading the complete sky. We mainly use it for the Default Sky since a full reload sets the sky scene using an animation. This method resizes without reloading.

Added Sky Dome Size Adjustment - We added an adjustment for the sky dome size when you select a Default Sky Type. The remainder of the current skies are boxes and use a different size adjustment on the form. The two size adjustments synchronize in code and feed the WTW.init.skySize global variable.

Increased the Max Camera View - We increased the camera maxZ to extend the view range from skybox size plus 1000 to skybox size plus 5000. When we only had dome skies, the 100 was enough to always be larger than the skydome. But now that we have sky boxes, the corners were getting clipped by the camera range on the larger skyboxes. This corrects that issue.