Updated 3.6.4
Changed order of Camera Loading - With the new SkyBoxes, some take longer to load causing the camera to not be loaded when the scene render loop begins. We now changed the order of loading to make the camera load before the sky, fog, and extended ground to make sure it completes before the render loop begins. This corrects the occasional no camera defined error message.